pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "halref"
version = "0.1.0"
description = "Detect hallucinated references in academic PDFs"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "typer>=0.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiohttp>=3.9",
    "aiolimiter>=1.1",
    "rapidfuzz>=3.0",
    "pdfminer.six>=20221105",
    "pypdf>=3.0",
    "pdfplumber>=0.11",
    "nameparser>=1.1",
    "bibtexparser>=1.4",
    "rich>=13.0",
]

[project.optional-dependencies]
docling = ["docling>=2.0"]
marker = ["marker-pdf>=1.0"]
llm = ["openai>=1.0"]
acl = ["acl-anthology>=0.6"]
all = ["docling>=2.0", "marker-pdf>=1.0", "openai>=1.0", "acl-anthology>=0.6"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "aioresponses>=0.7"]

[project.scripts]
halref = "halref.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
